separating user selection contexts into different component to allow access to other contexts
