+++ frontend/app/contexts/generalcontext.tsx
@@ -1,14 +1,13 @@
 "use client";
 import React, {createContext, Dispatch, useContext, useReducer} from 'react';
 import {allCensus, allQuadrats, Plot, plots} from "@/config/macros";
+import PlotProvider, {PlotsContext} from "@/app/contexts/userselectioncontext";
 
 export const PlotListContext = createContext<Plot[] | null>(null);
-export const PlotsContext = createContext<Plot | null>(null);
 export const CensusContext = createContext<number | null>(null);
 export const QuadratContext = createContext<number | null>(null);
 export const FirstLoadContext = createContext<boolean | null>(null);
 export const PlotListDispatchContext = createContext<Dispatch<{ plotList: Plot[] | null }> | null>(null);
-export const PlotsDispatchContext = createContext<Dispatch<{ plotKey: string | null }> | null>(null);
 export const CensusDispatchContext = createContext<Dispatch<{ census: number | null }> | null>(null);
 export const QuadratDispatchContext = createContext<Dispatch<{ quadrat: number | null }> | null>(null);
 export const FirstLoadDispatchContext = createContext<Dispatch<{ firstLoad: boolean }> | null>(null);
@@ -18,10 +17,6 @@ export function ContextsProvider({children}: { children: React.ReactNode }) {
     plotListReducer,
     plots
   )
-  const [plot, plotDispatch] = useReducer(
-    plotsReducer,
-    null
-  );
   const [census, censusDispatch] = useReducer(
     censusReducer,
     null
@@ -40,40 +35,25 @@ export function ContextsProvider({children}: { children: React.ReactNode }) {
     <>
       <PlotListContext.Provider value={plotList}>
         <PlotListDispatchContext.Provider value={plotListDispatch}>
-          <PlotsContext.Provider value={plot}>
-            <PlotsDispatchContext.Provider value={plotDispatch}>
-              <CensusContext.Provider value={census}>
-                <CensusDispatchContext.Provider value={censusDispatch}>
-                  <QuadratContext.Provider value={quadrat}>
-                    <QuadratDispatchContext.Provider value={quadratDispatch}>
-                      <FirstLoadContext.Provider value={firstLoad}>
-                        <FirstLoadDispatchContext.Provider value={firstLoadDispatch}>
-                          {children}
-                        </FirstLoadDispatchContext.Provider>
-                      </FirstLoadContext.Provider>
-                    </QuadratDispatchContext.Provider>
-                  </QuadratContext.Provider>
-                </CensusDispatchContext.Provider>
-              </CensusContext.Provider>
-            </PlotsDispatchContext.Provider>
-          </PlotsContext.Provider>
+          <PlotProvider>
+            <CensusContext.Provider value={census}>
+              <CensusDispatchContext.Provider value={censusDispatch}>
+                <QuadratContext.Provider value={quadrat}>
+                  <QuadratDispatchContext.Provider value={quadratDispatch}>
+                    <FirstLoadContext.Provider value={firstLoad}>
+                      <FirstLoadDispatchContext.Provider value={firstLoadDispatch}>
+                        {children}
+                      </FirstLoadDispatchContext.Provider>
+                    </FirstLoadContext.Provider>
+                  </QuadratDispatchContext.Provider>
+                </QuadratContext.Provider>
+              </CensusDispatchContext.Provider>
+            </CensusContext.Provider>
+          </PlotProvider>
         </PlotListDispatchContext.Provider>
       </PlotListContext.Provider>
     </>
   );
-}
-
-function plotsReducer(currentPlot: any, action: { plotKey: string | null }) {
-  const plotListContext = usePlotListContext();
-  if (plotListContext) {
-    if (action.plotKey == null) return null;
-    else if (plotListContext.find((p) => p.key == action.plotKey)) return plotListContext.find((p) => p.key == action.plotKey);
-    else return currentPlot;
-  } else {
-    if (action.plotKey == null) return null;
-    else if (plots.find((p) => p.key == action.plotKey)) return plots.find((p) => p.key == action.plotKey);
-    else return currentPlot;
-  }
 }
 
 function censusReducer(currentCensus: any, action: { census: number | null }) {
@@ -95,14 +75,6 @@ function firstLoadReducer(currentState: any, action: { firstLoad: boolean | null
 
 function plotListReducer(_currentPlotList: any, action: {plotList: Plot[] | null}) {
   return action.plotList;
-}
-
-export function usePlotContext() {
-  return useContext(PlotsContext);
-}
-
-export function usePlotDispatch() {
-  return useContext(PlotsDispatchContext);
 }
 
 export function useCensusContext() {
